backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.10.5
celery==5.4.0
redis==5.2.0
PyPDF2==3.0.1
cryptography==43.0.3
reportlab==4.2.5